refactor: change rpc references to chain
